Provide a filter for inventory-pickup event
This prevents the event from firing when there are no items to pick
up, thus enabling undoing moves that don't end on a hex containing
items. Secondly it also prevents this event from firing for
AI-controlled units. This is a sensible default, although it might
be a good idea to make it configurable.